fix: fix double free for SCStream
Fixes a double free issue when dropping an SCStream object. The issue
was caused because every time the SCStream was dropped, the Box stored in
StreamDelegateWrapper was also dropped. This could occur if the SCStream
object was cloned or when one of the SCStreamDelegateTrait functions
were called.

It is fixed here by doing reference count in the Cleanup object. This
requires to manually increase the reference count every time the
SCStream is increasing its reference count.